requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
python-dotenv
sqlalchemy
asyncpg
alembic
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose
email-validator